pyproject.toml
[tool.poetry]
name = "sound-cloud"
version = "0.0.1"
description = ""
authors = ["sultan <sultankurbanov404@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
black = "^24.3.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
